+++ Logos/Changes.md
@@ -17,6 +17,7 @@ Supports **Vault v2.1** and **Flow v2.8**
   - Added `.standardize(String)`
 ### Other
 - Moved statement-pulling functions from **AccessStatementPlacementRows** to **AccessStatementLinkRows**
+- `Domain.domainCharacterRegex` is now public
 
 ## v0.6 - 01.11.2025
 A small update following the changes in **Vault v2.0**.